fix(seeder): seed manager generate and seedString
Seeder manager now generates seed files with `export class` and the
correct seeder name (e.g. BookSeeder) instead of the default. Also when
seeding using string it executes the exported member instead
of the imported file itself.

Fixes #1843
